+++ project/src/components/app/app.tsx
@@ -1,4 +1,4 @@
-import Main, {FilmProps} from '../../pages/Main/Main';
+import Main, {FilmProps} from '../../pages/Main/main';
 
 function App(props: FilmProps) {
   return Main(props);
