+++ project/src/components/card_item.tsx
