+++ project/src/components/play-pause-button/pause_button.tsx
@@ -1,0 +1,10 @@
+export function RenderPauseButton() {
+  return (
+    <button type="button" className="player__play">
+      <svg viewBox="0 0 14 21" width="14" height="21">
+        <use xlinkHref="#pause"/>
+      </svg>
+      <span>Pause</span>
+    </button>
+  )
+}
