+++ project/src/components/play-pause-button/play_button.tsx
@@ -1,0 +1,10 @@
+export function RenderPlayButton() {
+  return (
+    <button type="button" className="player__play">
+      <svg viewBox="0 0 19 19" width="19" height="19">
+        <use xlinkHref="#play-s"/>
+      </svg>
+      <span>Play</span>
+    </button>
+  )
+}
