+++ project/src/pages/Main/main.tsx
@@ -1,6 +1,7 @@
-import Card from '../../components/CardItem';
+import Card from '../../components/card_item';
 import Films from '../../FilmsList';
-import Logo from '../../components/Logo';
+import Logo from '../../components/logo';
+import Footer from "../../components/footer";
 
 export interface FilmProps {
   filmName: string;
@@ -11,59 +12,7 @@ export interface FilmProps {
 function Main({filmName, filmGenre, filmDate}: FilmProps) {
   const films = new Films();
   return (
-    <body>
-      <div className='visually-hidden'>
-
-        <svg xmlns='http://www.w3.org/2000/svg' xmlnsXlink='http://www.w3.org/1999/xlink'>
-          <symbol id='add' viewBox='0 0 19 20'>
-            <title>+</title>
-            <desc>Created with Sketch.</desc>
-            <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-              <polygon id='+' fill='#EEE5B5'
-                points='10.777832 11.2880859 10.777832 19.5527344 8.41650391 19.5527344 8.41650391 11.2880859 0.627929688 11.2880859 0.627929688 8.92675781 8.41650391 8.92675781 8.41650391 0.662109375 10.777832 0.662109375 10.777832 8.92675781 18.5664062 8.92675781 18.5664062 11.2880859'
-              />
-            </g>
-          </symbol>
-          <symbol id='full-screen' viewBox='0 0 27 27'>
-            <path fillRule='evenodd' clipRule='evenodd' d='M23.8571 0H16V3.14286H23.8571V11H27V3.14286V0H23.8571Z'
-              fill='#FFF9D9' fillOpacity='0.7'
-            />
-            <path fillRule='evenodd' clipRule='evenodd' d='M27 23.8571V16H23.8571V23.8571H16V27H23.8571H27L27 23.8571Z'
-              fill='#FFF9D9' fillOpacity='0.7'
-            />
-            <path fillRule='evenodd' clipRule='evenodd'
-              d='M0 3.14286L0 11H3.14286L3.14286 3.14286L11 3.14286V0H3.14286H0L0 3.14286Z' fill='#FFF9D9'
-              fillOpacity='0.7'
-            />
-            <path fillRule='evenodd' clipRule='evenodd'
-              d='M3.14286 27H11V23.8571H3.14286L3.14286 16H0L0 23.8571V27H3.14286Z' fill='#FFF9D9' fillOpacity='0.7'
-            />
-          </symbol>
-          <symbol id='in-list' viewBox='0 0 18 14'>
-            <path fillRule='evenodd' clipRule='evenodd'
-              d='M2.40513 5.35353L6.1818 8.90902L15.5807 0L18 2.80485L6.18935 14L0 8.17346L2.40513 5.35353Z'
-              fill='#EEE5B5'
-            />
-          </symbol>
-          <symbol id='pause' viewBox='0 0 14 21'>
-            <symbol id='play-s' viewBox='0 0 19 19'>
-              <path fillRule='evenodd' clipRule='evenodd' d='M0 0L19 9.5L0 19V0Z' fill='#EEE5B5'/>
-            </symbol>
-
-            <title>Artboard</title>
-            <desc>Created with Sketch.</desc>
-            <g id='Artboard' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-              <polygon id='Line' fill='#EEE5B5' fillRule='nonzero'
-                points='0 -1.11910481e-13 4 -1.11910481e-13 4 21 0 21'
-              />
-              <polygon id='Line' fill='#EEE5B5' fillRule='nonzero'
-                points='10 -1.11910481e-13 14 -1.11910481e-13 14 21 10 21'
-              />
-            </g>
-          </symbol>
-        </svg>
-      </div>
-
+    <>
       <section className='film-card'>
         <div className='film-card__bg'>
           <img src='img/bg-the-grand-budapest-hotel.jpg' alt='The Grand Budapest Hotel'/>
@@ -89,7 +38,7 @@ function Main({filmName, filmGenre, filmDate}: FilmProps) {
           <div className='film-card__info'>
             <div className='film-card__poster'>
               <img src='img/the-grand-budapest-hotel-poster.jpg' alt='The Grand Budapest Hotel poster'
-                width='218' height='327'
+                   width='218' height='327'
               />
             </div>
 
@@ -159,12 +108,13 @@ function Main({filmName, filmGenre, filmDate}: FilmProps) {
 
           <div className='catalog__films-list'>
             <Card image={films.GrindelwaldImageData} href='film-page.html'
-              name='Fantastic Beasts: The Crimes of Grindelwald'
+                  name='Fantastic Beasts: The Crimes of Grindelwald'
             />
             <Card image={films.RhapsodyImageData} href='film-page.html' name='Bohemian Rhapsody'/>
             <Card image={films.MacbethImageData} href='film-page.html' name='Macbeth'/>
             <Card image={films.AviatorImageData} href='film-page.html' name='Aviator'/>
-            <Card image={films.WeNeedToTalkAboutKevinImageData} href='film-page.html' name='We need to talk about Kevin'/>
+            <Card image={films.WeNeedToTalkAboutKevinImageData} href='film-page.html'
+                  name='We need to talk about Kevin'/>
             <Card image={films.WhatWeDointheShadowsImageData} href='film-page.html' name='What We Do in the Shadows'/>
             <Card image={films.RevenantImageData} href='film-page.html' name='Revenant'/>
             <Card image={films.JohnnyEnglishImageData} href='film-page.html' name='Johnny English'/>
@@ -187,15 +137,9 @@ function Main({filmName, filmGenre, filmDate}: FilmProps) {
             <button className='catalog__button' type='button'>Show more</button>
           </div>
         </section>
-
-        <footer className='page-footer'>
-          <Logo className="logo__link logo__link--light"/>
-          <div className='copyright'>
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer/>
       </div>
-    </body>
+    </>
   );
 }
