+++ project/src/pages/sign-in/sign-in-error.tsx
@@ -1,0 +1,7 @@
+export function renderSignInError() {
+  return (
+    <div className="sign-in__message">
+      	<p>Please enter a valid email address</p>
+    </div>
+  );
+}
