+++ project/src/pages/sign-in/sign-in-message.tsx
@@ -1,0 +1,7 @@
+export function renderSignInMessage() {
+  return (
+    <div className="sign-in__message">
+      <p>We can’t recognize this email <br/> and password combination. Please try again.</p>
+    </div>
+  )
+}
